third_party/blink/web_tests/http/tests/xmlhttprequest/upload-request-error-event-order-expected.txt

CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/xmlhttprequest/resources/cross-site-progress-events.cgi' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource.
Test to validate the order in which the events are fired in case of a request error.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS actualOrder is "rsdone,upload.onerror,upload.onloadend,onerror,onloadend,"
PASS successfullyParsed is true

TEST COMPLETE
